Add files via upload
Pseudo-coded restricting moves due to check, and identifying checkmate. Have not yet done code for identifying check or some smaller functions needed for the aforementioned functions. Still need to change the pawn promotion to be a choice of piece instead of automatically selecting queen, and to include castling and en passant.