Add files via upload
Unbelievable progress was made on the game. The game is still lacking castling and en passant. There is also a bug where the king incorrectly has a destination removed from its list of possible destinations. This happens when an opposing piece would be threatening a destination space of the king but the opposing piece is blocked by another opposing piece that is not threatening the destination space. Essentially the game ignores the opposing piece that is blocking the otherwise threatening piece.
Other than these things, the game is very functional. It has some buttons and messages on the side. The Quit button does not work, but the Reset and Play Again buttons work.