Add files via upload
Castling working almost normally. Need to do more play testing. En passant not yet added. Buttons added for selecting a piece of choice for pawn promotion but the buttons are not yet functional (still automatically selects a queen). This should be a quick addition. Then the only remaining task is en passant.