Decision-making works more smoothly
Still makes blatantly bad decisions. Moves pieces of high value into direct line of pieces of lower value. The code is supposed to check for this to avoid it, but it does not work. To change: Remove the second layer of searching when moving pieces. Instead, search only for the optimal move at the first layer and perform additional checks at each potential destination.